Add team barriers.
At this point,
* KNL with >1 thread/team passes 'ctest -E r0 R ne2 -LE nightly' with
  8 threads/rank => 2 threads/team, but
* GPU fails with rel diffs at nearly machine precision because of the
  parallel_reduce calls in limiter_optim_iter_full.

This is the first time true nondeterminism has been introduced. Need
to think about whether we are willing to deal with the repercussions.

As a separate matter, right now I'm using stack-allocated arrays even
on GPU, which has the __shared__ issue: an array belongs to each
thread, so the algorithm is correct, but each thread has wasted stack
space. Next, I'm going to try an approach that uses scratch on GPU but
stack on CPU/KNL, abstracted with an exe-space template arg to a
container.