Memory::AutoArray
There are times when we want to use automatic (i.e., stack-allocated) arrays.

On GPU, this can work, but in a subtle way. Each thread on a warp gets its own
whole array and uses only the slots it indexes in a vector-direction
parallel_for. Other times, it doesn't work; that happens if one thread needs
data from another thread. In both cases, too much memory is used; most slots are
not used. The solution is to use scratch space.

Meanwhile, on CPU/KNL, each slot of the small automatic array is used, and auto
arrays (whether raw or std::array) are extremely efficient. We don't want to use
scratch space on CPU/KNL just because GPU needs it.

This class solves that problem.