Changed the logic of when the BE rebuilds its buffer views.
Instead of being the BM that asks the BE to rebuilds its internal views
when buffers (re)allocation happens, the BM now only clears the BE internal
views. It is up to the BE to then rebuild them when needed. In particular,
now the BE checks for its internal buffers to be valid at every call of
the exchange method. If they are not valid, it proceeds to (re)build them
(including the MPI requests). This reduces the number of times where the BM
explicitly uses its friendship with BE, and increases the control of the BE
structure over its own state.