Bugfix plus making a few lines shorter
The bug would probably never show up, but I believe was there. Now that we
allow multiple BE to use the same buffers, we MUST ensure that the send/recv
requests of a BE are completed before the next BE object can reuse the
buffers. This was already true for the recv requests (they must be completed
in order to unpack), but not for send requests (only the send requests of
the same BE object were tested, while the ones belonging to different BE's
cannot be reached).
Therefore, we now do a MPI_Waitall on the send requests at the end of the
call to exchange, to ensure that all the buffers are reusable upon the
return of the method.