Adding a feature to the boundary exchange structure.
The BE can now perform the neighbor min/max exchange, where 2 quantities
(min and max of some field) are exchange per element-level. This requires
different pack/unpack routines, since only 2 real per level are echanged
with each neighbor, opposed with the 'usual' echange, where 1 real per GP
is echanged with each neighbor, and in addition the exchanged values are
not accumulated into the local ones, but the local is set to the max/min
between local and neighbor's.

Due to these differences, we prohibit to use the same BE structure for both
'normal' and 'min/max' exhcange. Insteado, one should use two separate BE
structures. Notice, however, that the same BM structure can still be used
to serve both kind of BE at the same time.

We also added some more checks on BE, to make sure there is no send/recv
pending when there should be none.

Now, when passing the # of fields to the BE/BM objects, one has to pass all
three of them: 1d, 2d and 3d. In the BE we don't yet check, at this point,
that no 'mixed' use is requested (with both 1d and 2d/3d fields). However,
when registering fields, we make sure that, if 1d fields are registered,
no 2d/3d fields can be registered, and viceversa. This ensures that no user
makes a 'mixed' use of a single BE structure.