Adapting BE to perform exchanges over a given subset of the elements on proc.
Not that it probably will never be used, but it's more consistent with the
other C++ kernels. Also, I added a double check to make sure the connectivity
in the BE and the one in the BM are the same. Ideally, we should not be able
to set both of them, but I don't want to dwell on a cute design right now
(deadlines).