Added a way to use the BM to make sure buffers are available
We can in theory use interleaved pack/unpack method from different BE objects.
It would be wrong, but it is allowed. Therefore, we need to check whether
the buffers are busy (due to another BE using them). The right way to do it
is to add lock/unlock/query methods to BM.