Bugfix in BoundaryExchange.
When running with multiple BE, in the case a BE forces reallocation of
the BM internal buffers, we clear the buffer views in the other BE's. This
was leaving us with null buffer views in other BE, which were not
reallocated anymore. The only place where buffer views in BE were allocated
(but not filled) was in set_num_fields. Instead, now we allocate those
views right before filling them, in the build_buffer_views_and_requests,
so that we are SURE the views are valid when we fill them, even if they
were destroyed by a call to clear_buffer_views_and_requests from the BM.