Started the implementation of C-stack for prim routines
The new files aim to have a single C stack of function, starting high up
the stack (prim_main.F90), without the need to copy back data to F90.

Perhaps we should make a single big class, with these routines as method,
so we avoid abusing of the Context singleton.