Adjusted interfaces between C and F90.
Now all the interfaces are in the setup phase (in prim_init1/2), and in
the main.

I have removed ALL the code that depends on #ifdef USE_KOKKOS_KERNELS from
the prim_*_mod.F90 modules, since everything is now handled from the
prim_driver.cpp file (and functions called therein).