Tweak ride nightly.
It's not clear to me what's happening with the ride nightly. The nightly output
looks innocuous, but it simply stops midway.

To diagnose this, I'm temporarily excluding the "nightly" label and bringing the
timeout down slightly. This should make the tests finish pretty quickly. We'll
see if we get ride results tmw morning.