Speed up pack/unpack; CH in kernel was main culprit.
Stack-allocated ConnectionHelpers was blowing up the pack time on GPU. Copies
rather than refs contributed a little to the pack/unpack slowdown, too. This
commit fixes those issues.

I'm leaving a bunch of timers in for a little while. We'll strip them in a week
or two. But I want to make sure the MPI situation on GPU stabilizes first.