Bug fix in advance_hypervis_dp_c
We were not forwarding np1 to the control structure. Don't be fooled by the
fact that we do not touch np1 between the rk5 stages and the HVF call. The
timelevels are update in CaarFunctor::set_rk_stage_data, but the functor
only stores a COPY of the data, so the singleton's Control is NOT changed.
Which is why this commit is necessary.