Reducing C++ build compilation time by excluding some (long) F90 files
Some other upstream files required some doctoring, to exclude some function
calls in case of C++ build. This should be harmless, since these function
calls happened in configuraiton not supported by our test system (so far).
To be safe, in case those functions are called in a C++ build, we abort
with a message explaining this configuration is not supported.