Preqx build fix, following removal of F90 files from C++ build
This is puzzling: prtcA_flat_c built, but prtcB_flat_c did not build.
Whatever is the reason, this commit fixes the problem, making one use
directive conditional to CAM being defined (as it should), and removing
one use directive that brough in scope a subroutine that was not even
used.