Removing no longer present C interfaces form F90 files.
Also, some code will never be used in a CXX build, so it made sense to
guard it with #ifndef USE_KOKKOS_KERNELS. This is not an exaustive attempt
to reduce the code compiled in a CXX build to its minimum. But while I
was at it, it made sense to exclude the obvious parts.