Bug fix in prim_driver.cpp
The value of nete should not be decrease by one, since C ranges are open
on the right (while F ranges are closes). But in any case, we are changing
the prim_run_subcycle_c interface to not take nets/nete anymore. In C, we
always assume nets=0, and nete equal to the # of elements on this rank. In
other words, we don't allow C functions to be called from a horizontally
threaded region in Fortran.