KokkosKernel stuff: changing enum into static const int for vector_length
On one hand, this is how it is currently implemented in trilinos' master,
so it should be 'safe'. On the other hand, we were experiencing a mysterious
bug on the cuda build on white, where the vector_length enum in the SIMD
vector was 0, even though it was init-ed with tag_type::length, which was
NOT 0. I tried to google a bit, and even tried to generate a short cpp
file to reproduce the behavior, with no luck. The logic of getting
the vector_length as an enum, and init-ing it with tag_type::length
seemed legit, and I can't see why it would not work. Still, it didn't.

Worth noticing it, that also doing
   enum : int {
     vector_length = static_cast<int>(tag_type::length)
   };
worked. However, given the direction taken by trilinos' version of
kokkos-kernels, I opted for switching to static const int in our code
as well.