Making the nameslist in the test_exec/prtc* folders usable
Before, they contained two non-expanded cmake variable, HOMME_TESTS_NE
and HOMME_TEST_NDAYS. As such, they could not be used as inputs directly,
but had to be doctored first.

Now the two variable are expanded to HOMME_TESTS_NE=2 and HOMME_TESTS_NDAYS=1,
which correspond to what the 'dev' label would do. The idea is that those
namelists are for convenience, to quickly test the executable during the
development/debug stage, so smaller simulations are convenient.