Changed position of dvv argument in sphere operators
This is mostly 'style', but to me it is easier to read the code when
all the 'Homme::subview' arguments are one right below each other.
Conceptually, these are space fields, while dvv is just a math tensor,
so I'd like it not to be mixed with the others, and either come first
or last.

Also, I changed preqx_flat_ut_sphere_op_*l.cpp to just create a subview
inside the function call, rather than have a (somewhat hard to read)
declaration of the temporary. This helped reducing the # of lines of that
file too.

Finally, I also removed, where possible and appropriate, the use of
Kokkos::subview, in favor of Homme::subview, which is the optimized
version for LayoutRight.