ESF: Always run separate setup and Q phases.
We care most about the large-qsize case. In pratice qsize=40 is our target, but
I consider qsize >= 20 to be large. In that regime, it makes sense always to pay
for two OpenMP dispatches rather than attempt to fuse kernels. Do this in
ESF. This has a nice positive impact on ESF speed on KNL for qsize >= 20, but
it's a bit of a hit for qsize ~ <= 10. In principle, we could switch on qsize,
but until we study this more, I prefer to follow a simple rule of dispatching
twice whenever there are Q-independent and Q-dependent phases.