Random init-ing ps_v and eta_dot_dpdn in a more robust way.
The two need to have certain magnitudes to allow remap_ut to not fail when
dt>0.

In particular, if dt>0, we need
   dp3d[k] + dt*(eta_dot_dpdn[k+1]-eta_dot_dpdn[k]) > 0.

To this end, we make ps_v 2-3 orders of magnitude larger than ps0, and we
make eta_dot_dpdn ~ 1-2 orders of magnitude smaller than dp3d. This allows
us (read: *should* allow us) to use dt in [0.01,10] without crashing.