ESF: Convert some RangePolicy kernels to TeamPolicy. Microopt in limiter.
On CPU/KNL, there's reason to think RangePolicy-based kernels with (1) a lot of
div-mod arithmetic and (2) full array indexing relative to actual work might be
slower than we'd like. TeamPolicy-based kernels avoid both.

In BoundaryExchange, we haven't yet gotten the same conversion to be faster on
GPU, too. But in this commit, the conversions give speedup on all platforms.