Fixing the previous bug fix.
Even if eta_dot_dpdn_buf was always used, its use was incorrect. Now that
we rely on KernelsBuffersManager, the piece of buffer for eta_dot_dpdn_buf
was also used by other parts of the code (outside CaarFunctorImpl), which
led to it not being identically zero. The two solutions were

1) zero out eta_dot_dpdn before each call to CaarFunctorImpl (or, at lest,
   before the first RK stage)
2) use an if statement, and only use eta_dot_dpdn_buf if rsplit=0.

I opted for the second, which is cleaner, and also agreed upon by other
developers.