Changing the way we handle buffers on GPU
Unfortunately, the nice and fancy way of getting the number
of concurrent teams does not work on GPU, due to some issues
in getting the code to compile (blockIdx & co. cannot appear
in a __host__ function, and I could not find a way to get the
code to compile on GPU). On GPU we therefore still use
num_elems*qsize as number of buffers, and we use ie,iq to
access the correct buffer. To uniform the code in SphereOperators,
we still use kv.team_idx to access the buffer, but that index
is now computed in terms of ie,qsize,iq, rather than from
thread id's information.