Modifying BE registration methods, to allow registration of single-element fields.
This required to add some internal counters, and to remove parallel for's
from the registration methods (which, in turn, requires to carefully use
host mirrors for the registered fields views of views). This is not a heavy
burden, as it happens only at setup time, and there was not a whole lot of
concurrency to start with.