Change in how fields are stored in BoundaryExchange.
Before, the VoV (view of views) was managed of managed. Now we do a managed
view of unmanaged views, because the fields we register may not be managed,
and I think this may generate ownership problems (double delete or dangling
pointer, or something along that line).

Also, since the subviews are unmanaged, we can now use Homme::subview.