Alloc qmin/max only ifndef USE_KOKKOS_KERNELS.
The next three commits reduce memory usage due to tracers.

Before this commit, in principle the C++ code should be able to run QSIZE_D=40,
ne=15, but not ne=16, on 1 KNL node and stay in HBM, so up to ~1350
elem/node. One can track this via numactl -H.

For ne=15, QSIZE_D=40, this commit reduces memory use on a node by ~1.4 GB.

The next commit, for vstar_qdp, reduces it by another ~0.9 GB.

The one after that, for qtens, reduces it by another ~0.45 GB, for a total
reduction of just under 3 GB.