Fully vectorize serial_limiter_optim_iter_full.
This commit is a slowdown on HSW, despite getting serial_limiter_optim_iter_full
to vectorize fully. That is because full vectorization is too work
inefficient. Thus, I'm pushing this commit for purposes of documentation in the
git history, but will then immediatley revert it.

The speedup on Edison comes from the key fact that the limiter in the not fully
unvectorized form is only slightly work inefficient but acesses memory
essentially optimally. On older architectures, memory access is relatively more
important.