Added test for diagnostics
Unless disable_diagnostics=true, all runs print diagnostics at least at the
last time step. So we add another set of comparison to the cxx_vs_f90 tests.
These F-C tests grep the raw output of the F and C tests, looking
for diagnostics lines (as of today, they are all and only the lines containing
the string '/dt'), and compare the grepped lines. If matching, the two tests
compute the same diagnostics.