Forcing: Some bug fixes.
Fix the unit test for corrected tl fields.

Move reduction for ps_v to before the qdp update, so that v1 is correct in the
case that clipping is applied.

Use kv.ie, not team_idx, to get ie.

Fix some dt stuff, and make dt something other than 1 in the unit test to test
correctness wrt dt_q.